+++ CHANGELOG.rst
@@ -3,15 +3,15 @@
 Changelog
 ---------
 
-1.2.1
-+++++
+1.2.1 (2015-04-03)
+++++++++++++++++++
 
 - Rename ``put`` to ``set`` in ``redisext.hashmap``
 - Method ``exists`` method to ``Map``
 - Connections re-use
 - Add ``Lock``
 
-1.2.0 (2015-04-03)
+1.2.0 (2015-04-02)
 ++++++++++++++++++
 
 - Connection object uses through CONNECTION attribute, instead of inheritance